+++ drivers/interrupt_controller/loapic_intr.c
@@ -17,9 +17,9 @@
 
 #include <toolchain.h>
 #include <linker/sections.h>
-#include <drivers/loapic.h> /* public API declarations */
+#include <drivers/interrupt_controller/loapic.h> /* public API declarations */
 #include <init.h>
-#include <drivers/sysapic.h>
+#include <drivers/interrupt_controller/sysapic.h>
 
 /* Local APIC Version Register Bits */
